fix(graphics): 🐛 Fixed an issue with the graphics focussing window and also some minor possible bug with loading.py

+++ graphics/loading.py
@@ -30,10 +30,10 @@ class thread_with_exception(Thread):
   
     def raise_exception(self):
         thread_id = self.get_id()
-        res = ctypes.pythonapi.PyThreadState_SetAsyncExc(thread_id,
+        res = ctypes.pythonapi.PyThreadState_SetAsyncExc(ctypes.c_long(thread_id),
               ctypes.py_object(SystemExit))
         if res > 1:
-            ctypes.pythonapi.PyThreadState_SetAsyncExc(thread_id, 0)
+            ctypes.pythonapi.PyThreadState_SetAsyncExc(ctypes.c_long(thread_id), 0)
             print('Exception raise failure')
 
 class LoadingScreen:
